Refactor UI directory structure
