+++ mobile/lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:mobile/core/core.dart';
-import 'package:mobile/ui/theme/theme.dart';
+import 'package:mobile/ui/theme/app_theme.dart';
 
 Future<void> main() async {
   await dotenv.load();
