+++ mobile/lib/ui/authentication/authentication.dart
@@ -1,0 +1,1 @@
+export 'pages/pages.dart';
