+++ mobile/lib/ui/authentication/pages/pages.dart
@@ -1,0 +1,2 @@
+export 'sign_in_page.dart';
+export 'sign_up_page.dart';
