+++ mobile/lib/ui/home/home.dart
@@ -1,0 +1,1 @@
+export 'pages/pages.dart';
