+++ mobile/lib/ui/splash/pages/pages.dart
@@ -1,0 +1,1 @@
+export 'splash_page.dart';
