+++ mobile/lib/ui/splash/splash.dart
@@ -1,0 +1,1 @@
+export 'pages/pages.dart';
