+++ mobile/lib/ui/ui.dart
@@ -1,0 +1,4 @@
+export 'authentication/authentication.dart';
+export 'home/home.dart';
+export 'splash/splash.dart';
+export 'theme/theme.dart';
